requirements.txt
streamlit
langchain-openai
python-dotenv
scikit-learn